+++ packages/frameworks/src/frameworks.ts
@@ -4730,7 +4730,7 @@ export const frameworks = [
     name: 'Services',
     slug: 'services',
     experimental: true,
-    logo: 'https://api-frameworks.vercel.sh/framework-logos/other.svg',
+    logo: 'https://api-frameworks.vercel.sh/framework-logos/services.svg',
     tagline:
       'Multiple services deployed as serverless functions within your project.',
     description: