+++ packages/frameworks/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @vercel/frameworks
 
+## 3.30.4
+
+### Patch Changes
+
+- d19dbc5: Use circle icons for dark mode and light mode for services and containers
+
+## 3.30.3
+
+### Patch Changes
+
+- 02d8bab: Add services logo
+
+## 3.30.2
+
+### Patch Changes
+
+- acddddf: Update the Container framework logo to use the Box icon.
+
 ## 3.30.1
 
 ### Patch Changes